+++ lib/basic.go
@@ -1,4 +1,4 @@
-package cinnabot
+package lib
 
 import (
 	"strings"
