+++ lib/basic_test.go
@@ -1,4 +1,4 @@
-package cinnabot
+package lib
 
 import (
 	"os"
